refactor(MatchResult): enum 클래스명 변경, enum field 추가
