+++ src/main/java/model/participant/dto/PlayerMatchResult.java
@@ -1,9 +1,9 @@
 package model.participant.dto;
 
-import model.participant.MatchResult;
+import model.participant.MatchState;
 import model.participant.Name;
 
-public record PlayerMatchResult(Name name, MatchResult matchResult) {
+public record PlayerMatchResult(Name name, MatchState matchState) {
     public String getNameAsString() {
         return name.getValue();
     }
